Correct working set of stiffness matrix
The stiffness matrix computation proceeds as Felix did in his matlab
code. This looks correct after modifying the boundary conditon
application that had a factor of -1 missing that arises out of the
constituive equaition described in 5.2 of roadmap to software
implementation, handbook of peridynamic modeling

TODO:
1. extend the code to generalized case of boundary condtion
applicatioapplication, mesh with holes and 3d meshes.

2. Optimize the code (possibly parallelize it too)

3. Not necessary but for fun: try 4D and 5D meshes too